third_party/blink/web_tests/platform/generic/fast/parser/fragment-parser-doctype-expected.txt

Verify that a fragment's DOCTYPE does not affect parsing. We expect DOCTYPEs to be ignored for fragments with context elements.
PASS container.firstChild.nextSibling is null
Verify that a fragment's DOCTYPE does not change the compatibility mode of the owner document.
PASS container.firstChild.nextSibling is null
PASS successfullyParsed is true

TEST COMPLETE
